change eit for gcr
